[Fix](tablet id) use int64_t instead of int32_t or uint32_t for tablet_id (#34962)

+++ be/src/olap/rowset/segment_v2/column_reader.h
@@ -550,7 +550,7 @@ private:
 class RowIdColumnIterator : public ColumnIterator {
 public:
     RowIdColumnIterator() = delete;
-    RowIdColumnIterator(int32_t tid, RowsetId rid, int32_t segid)
+    RowIdColumnIterator(int64_t tid, RowsetId rid, int32_t segid)
             : _tablet_id(tid), _rowset_id(rid), _segment_id(segid) {}
 
     Status seek_to_first() override {
